+++ app/role/role.go
@@ -13,10 +13,10 @@ type Role struct {
 
 // RoleInterface defines the set of methods that any implementation of the Role service must provide.
 type RoleInterface interface {
-	CreateRole(ctx context.Context, role Role) (int64, error) // Creates a new role and returns the role ID
-	GetRoleByID(ctx context.Context, id int64) (*Role, error) // Fetches a role by their ID
+	CreateRole(ctx context.Context, role Role) (int, error) // Creates a new role and returns the role ID
+	GetRoleByID(ctx context.Context, id int) (*Role, error) // Fetches a role by their ID
 	UpdateRole(ctx context.Context, role Role) error          // Updates an existing role
-	DeleteRole(ctx context.Context, id int64) error           // Deletes a role by their ID
+	DeleteRole(ctx context.Context, id int) error           // Deletes a role by their ID
 	ListRoles(ctx context.Context, language string) (*[]RoleEntity, error)
 }
 
@@ -26,7 +26,7 @@ type NewRoleRequest struct {
 }
 
 type UpdateRoleRequest struct {
-	ID   uint   `json:"id"`
+	ID   int   `json:"id"`
 	Name string `json:"name"`
 	Language  string  `json:"language"`
 }
